API improvements plus performance and testing.
API improvements include the addition of `Fidelity.resolve()` and
`Fidelity.deferred()`. See the documentation changes for more details.

This commit fixes #1 with the addition of performance benchmarks and a
new build target, `npm run benchmarks`. As I was putting together the
benchmarks for this commit, I made a number of changes to the code to
speed up performance. These included using `const` and arrow functions
where possible, as well as inlining a number of helper functions.

I've also added a bunch of build targets and some structure around CI.
New build targets are:

    $ make ci
    $ make test
    $ make list
    $ make clean
    $ npm run docs
    $ npm run benchmarks
    $ npm run profile
    $ npm run coverage
    $ npm run prepublish

This last command simply profiles the small number of unit tests in the
`tape-test.js` test script. This should be improved to include the
`promises-aplus-tests`. The coverage target should include this as well.

Bump version number

Set up publication of API docs.